refactoring: extract actor into own file
