+++ src/actor/mod.rs
@@ -9,6 +9,7 @@ pub mod context;
 pub mod executor;
 pub mod handler;
 pub mod mailbox;
+pub mod actor;
 
 pub mod prelude {
     pub use crate::actor::actor_config::RestartPolicy;
@@ -17,5 +18,5 @@ pub mod prelude {
     pub use crate::actor::actor_factory::ActorFactory;
     pub use crate::actor::context::ActorContext;
     pub use crate::actor::handler::Handler;
-    pub use crate::actor::handler::Actor;
+    pub use crate::actor::actor::Actor;
 }
